changed

+++ app/Http/Controllers/API/v1/Dashboard/Seller/ShopCategoryController.php
@@ -31,10 +31,14 @@ class ShopCategoryController extends SellerBaseController
      * Display a listing of the resource.
      *
      * @param CategoryFilterRequest $request
-     * @return AnonymousResourceCollection
+     * @return JsonResponse|AnonymousResourceCollection
      */
-    public function index(CategoryFilterRequest $request): AnonymousResourceCollection
+    public function index(CategoryFilterRequest $request)
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $shopBrands = $this->categoryRepository->shopCategory($request->merge(['shop_id' => $this->shop->id])->all());
 
         return CategoryResource::collection($shopBrands);
@@ -48,6 +52,10 @@ class ShopCategoryController extends SellerBaseController
      */
     public function store(StoreRequest $request): JsonResponse
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $validated = $request->validated();
         $validated['shop_id'] = $this->shop->id;
 
@@ -57,10 +65,7 @@ class ShopCategoryController extends SellerBaseController
             return $this->onErrorResponse($result);
         }
 
-        return $this->successResponse(
-            __('errors.' . ResponseError::RECORD_WAS_SUCCESSFULLY_CREATED, locale: $this->language),
-            data_get($result, 'data')
-        );
+        return $this->successResponse(__('web.record_successfully_created'), data_get($result, 'data'));
     }
 
     /**
@@ -71,19 +76,17 @@ class ShopCategoryController extends SellerBaseController
      */
     public function show(int $id): JsonResponse
     {
-        $model = $this->categoryRepository->shopCategoryById($id, $this->shop->id);
-
-        if (!$model) {
-            return $this->onErrorResponse([
-                'code'      => ResponseError::ERROR_404,
-                'message'   => __('errors.' . ResponseError::ERROR_404, locale: $this->language)
-            ]);
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
         }
 
-        return $this->successResponse(
-            __('errors.' . ResponseError::SUCCESS, locale: $this->language),
-            CategoryResource::make($model)
-        );
+        $shopBrand = $this->categoryRepository->shopCategoryById($id, $this->shop->id);
+
+        if (!$shopBrand) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_404]);
+        }
+
+        return $this->successResponse(__('web.coupon_found'), CategoryResource::make($shopBrand));
     }
 
     /**
@@ -94,6 +97,10 @@ class ShopCategoryController extends SellerBaseController
      */
     public function update(UpdateRequest $request): JsonResponse
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $collection = $request->validated();
         $collection['shop_id'] = $this->shop->id;
 
@@ -103,10 +110,7 @@ class ShopCategoryController extends SellerBaseController
             return $this->onErrorResponse($result);
         }
 
-        return $this->successResponse(
-            __('errors.' . ResponseError::RECORD_WAS_SUCCESSFULLY_UPDATED, locale: $this->language),
-            data_get($result, 'data')
-        );
+        return $this->successResponse(__('web.record_has_been_successfully_updated'), data_get($result, 'data'));
     }
 
     /**
@@ -117,19 +121,25 @@ class ShopCategoryController extends SellerBaseController
      */
     public function destroy(FilterParamsRequest $request): JsonResponse
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $this->shopCategoryService->delete($request->input('ids', []), $this->shop->id);
 
-        return $this->successResponse(
-            __('errors.' . ResponseError::RECORD_WAS_SUCCESSFULLY_DELETED, locale: $this->language)
-        );
+        return $this->successResponse(__('web.record_has_been_successfully_delete'));
     }
 
     /**
      * @param CategoryFilterRequest $request
-     * @return AnonymousResourceCollection
+     * @return JsonResponse|AnonymousResourceCollection
      */
-    public function allCategory(CategoryFilterRequest $request): AnonymousResourceCollection
+    public function allCategory(CategoryFilterRequest $request)
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $category = $this->categoryRepository->shopCategoryNonExistPaginate(
             $request->merge(['shop_id' => $this->shop->id])->all()
         );
@@ -141,10 +151,14 @@ class ShopCategoryController extends SellerBaseController
      * Display a listing of the resource.
      *
      * @param CategoryFilterRequest $request
-     * @return AnonymousResourceCollection
+     * @return JsonResponse|AnonymousResourceCollection
      */
-    public function paginate(CategoryFilterRequest $request): AnonymousResourceCollection
+    public function paginate(CategoryFilterRequest $request)
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $shopBrands = $this->categoryRepository->shopCategoryPaginate(
             $request->merge(['shop_id' => $this->shop->id])->all()
         );
@@ -156,10 +170,14 @@ class ShopCategoryController extends SellerBaseController
      * Display a listing of the resource.
      *
      * @param CategoryFilterRequest $request
-     * @return AnonymousResourceCollection
+     * @return JsonResponse|AnonymousResourceCollection
      */
-    public function selectPaginate(CategoryFilterRequest $request): AnonymousResourceCollection
+    public function selectPaginate(CategoryFilterRequest $request)
     {
+        if (!$this->shop) {
+            return $this->onErrorResponse(['code' => ResponseError::ERROR_204]);
+        }
+
         $shopBrands = $this->categoryRepository->selectPaginate(
             $request->merge(['shop_id' => $this->shop->id, 'type' => Category::MAIN])->all()
         );
